refactor: restructure clients layer with directory organization
- Reorganize aizu_online_judge and atcoder_problems into subdirectories
- Split AOJ client into clients.ts (API), types.ts (types), utils.ts (helpers)
- Move test data to unified fixtures/ directory
- Extract HTTP client helpers to fixtures/helpers.ts
- Update imports across codebase to reflect new structure
- Add comprehensive plan for clients restructuring (Phase 1 complete)
